Update README.md
I created this Morse Code Converter as a personal challenge to get better at Python and figure out how to build a GUI with Tkinter. I’ve always thought Morse code was cool because of its history, so I decided to turn it into a project. It started small, but I kept adding stuff like the copy feature to make it more useful. Figuring out how to handle spaces and optimize the code took some work, but I’m happy with how it turned out. Hope you find it fun to play with!